fix: image storage and retrieval for university proof
The storage configuration for university images has been moved to `users.controller.ts`.
This refactor better organizes and centralizes the handling of these images within the code.
Additionally, the user service has been updated to use the image filename rather than path.
