+++ libs/sugaming-services/src/users/users.service.ts
@@ -579,7 +579,7 @@ export class UsersService {
         passwordHash: hashedPassword,
         universityProofImages: {
           create: universityProofImages.map((image) => ({
-            url: image.path,
+            url: image.filename,
           })),
         },
         ...userToBeCreatedWithoutPassword,
